testutil/beaconmock: add static provider (#185)
Adds a static values providing beacon mock. This will be required for testing the SigAgg component since it depends on the eth2client.DomainProvider. It adds real world static value to the beaconmock by serving an embedded json file. A script is included to generate the json.

category: testing
ticket: #184