cmd: refactor create-cluster flags (#385)
Refactors `create-cluster` command flags to improve UX when used in combination with docker-compose and k8s.

The main purpose of `create-cluster` is to create a manifest and keys. Creating **config run scripts (run.sh)** is only applicable to local non-docker clusters (which will not be the most common use case). So make creating config run scripts opt-in by specifying `--config=true`. This also makes it more explicit that configuring a `--config-simnet` is just via the config run script. The same goes for `--config-port-start` and `--config-binary`. 

When used in charon-docker-compose or charon-k8s, this config is not applicable and is now not generated any more. So less confusing and doesn't need to be deleted.

Also align `--split-keys-dir` with `--split-existing-keys` with same prefix to indicate they are related. 

category: refactor 
ticket: none