build: charon cluster deployment workflow (#723)
Add a step to the build workflow to trigger/dispatch the Charon cluster deployment workflow in the Charon-k8s repository. This will deploy/update the long-lived Charon clusters running on the GKE dev env.

This step dispatches a package published event each time a new Charon image is pushed to the Github registry. The actual deployment workflow is in the charon-k8s repo.


category: misc
ticket: none