core/scheduler: workaround clock skew issue (#765)
This is a workaround for sporadic clock skew wrt beacon node. Instead of querying for what we think is the current slot, just query the "head". 

Note this does introduce a small risk of off-by-one issues in the "active validator logic". But this is only applicable to when validators become active or inactive and has insignificant impact.  

category: bug 
ticket: #764