dockerfile: try and collide docker uid+gid with host uid+gid to reduce permission denied errors (#780)
Right now charon's user ID is 12345 and group ID is 23456, neither of which are ever present on a host. Instead, defaulting to 1000:1000 seems to be what unix defaults its first user account to. That way, although charon still not running as root, it should have the ability to write to the host folder without it, provided the linux user is using the first user on their machine (ids start at 1000 and work up)

category: bug
ticket: none