app/errors: add sentinel error without stacktrace (#1483)
Formalises the concept of a sentinel error. Package level global errors that are "catchable". They should always be wrapped when returned which adds the proper stack trace. 

This is required to fix a bug in `inclusion tracker` where scheduler returns a error if the duty is not found so we need to catch it.

category: misc
ticket: none