dkg: partial deposits enabled (#2894)
* `dkg` command supporting partial deposits (previous work added cli option `--deposit-amounts` to `create dkg` command).

One important and _not very elegant_ change was done to the `exchanger` component, specifically:
```
	// sigDepositData is responsible for deposit data signed partial signatures exchange and aggregation.
	// For partial deposits, it increments the number for each unique partial amount, e.g. 201, 202, etc.
	sigDepositData sigType = 200
```

This is because we need to execute several rounds for each partial deposit amount and we need to distinguish between each such round and store them independently (parsigdb). To make it more elegant, a drastic change to the protocol would be required, such as adding additional field to the tuple of {Duty, Slot}, I would call it a group. But when I tried to go this route, I quickly realized the viral effect of this change and too many components will be affected. Hence the simplification and the present solution. Open to discuss alternative options.

category: feature
ticket: #2889