*: check for attestation inclusion only under feature flag (#3740)
Previously was checked only the slot of the first attestation in a block, which caused some attestations submitted wrongly deemed as not included on-chain.

~TODO: Still to be tested on a Hoodi cluster before merging.~
After testing on Hoodi this seems to add A LOT of load on the beacon node. I have introduced caching of the state, but even that is putting a lot of stress. Calling for state on each slot is expensive. Sometimes those requests timeout and we again fail to check the attestation, as we don't have the data from the BN.

~I personally am in favour of removing the attestations for the on-chain inclusion tracking or delegating the whole on-chain inclusion tracking to a separate BN.~

It has been agreed to change the default scenario to not tracking attestations and aggregations, but still provide a feature flag for people that are willing to put the BN under high load to track those.

category: feature
ticket: none
feature_flag: attestation_inclusion